+++ src/main/java/ss/parser/ad/flat/Flat.java
@@ -16,10 +16,11 @@ class Flat extends AbstractAd {
     private final int floor;
     private final int floors;
     private final int price;
-    private final int ppm2;
+    /** Price per m², a decimal because rent feeds quote values such as 7.76 €. */
+    private final double ppm2;
 
     Flat(Element element, String region, String address, String series,
-         int rooms, int area, int floor, int floors, int price, int ppm2) {
+         int rooms, int area, int floor, int floors, int price, double ppm2) {
         super(element);
         this.region = region;
         this.address = address;
